+++ app/src/main/java/com/example/swith/data/Session.kt
@@ -1,13 +1,13 @@
 package com.example.swith.data
 
 data class Session(
-    val groupIdx: Int,
+    val groupIdx: Long,
     val online: Int,
     val place: String,
     val sessionContent: String,
     val sessionEnd: String,
     val sessionStart: String,
-    val userIdx: Int
+    val userIdx: Long
 )
 
 data class SessionCreate(
@@ -29,6 +29,6 @@ data class SessionModify(
     val place: String,
     val sessionContent: String,
     val sessionEnd: String,
-    val sessionIdx: Int,
+    val sessionIdx: Long,
     val sessionStart: String
 )
