+++ app/src/main/java/com/example/swith/repository/round/update/RoundUpdateRepository.kt
@@ -8,8 +8,8 @@ import com.example.swith.utils.error.RemoteErrorEmitter
 
 
 class RoundUpdateRepository(private val roundUpdateRemoteDataSource: RoundUpdateRemoteDataSource) {
-    suspend fun getPostRound(errorEmitter: RemoteErrorEmitter, userIdx: Int ,groupIdx: Int) = roundUpdateRemoteDataSource.getPostRound(errorEmitter, userIdx, groupIdx)
+    suspend fun getPostRound(errorEmitter: RemoteErrorEmitter, userIdx: Long ,groupIdx: Long) = roundUpdateRemoteDataSource.getPostRound(errorEmitter, userIdx, groupIdx)
     suspend fun createRound(remoteErrorEmitter: RemoteErrorEmitter, session: Session) = roundUpdateRemoteDataSource.createRound(remoteErrorEmitter, session)
-    suspend fun deleteRound(remoteErrorEmitter: RemoteErrorEmitter, sessionIdx: Int) = roundUpdateRemoteDataSource.deleteRound(remoteErrorEmitter, sessionIdx)
+    suspend fun deleteRound(remoteErrorEmitter: RemoteErrorEmitter, sessionIdx: Long) = roundUpdateRemoteDataSource.deleteRound(remoteErrorEmitter, sessionIdx)
     suspend fun modifyRound(remoteErrorEmitter: RemoteErrorEmitter, session: SessionModify) = roundUpdateRemoteDataSource.modifyRound(remoteErrorEmitter, session)
 }
